requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.0
filelock==3.10.0
huggingface-hub==0.13.2
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lit==15.0.7
llvmlite==0.40.1
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
nltk==3.8.1
numba==0.57.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.15.0
pynndescent==0.5.10
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.1
triton==2.0.0
typing_extensions==4.5.0
umap-learn==0.5.3
urllib3==1.26.15
